requirements.txt
telethon==1.36.0
psycopg2-binary==2.9.9
requests==2.32.3
google-generativeai==0.7.2
python-dotenv==1.0.1
celery==5.3.4
redis==4.6.0
pymongo==4.5.0
